fix: tests for location
