+++ internal/cmd/locations_utils.go
@@ -1,6 +1,8 @@
 package cmd
 
-import megaport "github.com/megaport/megaportgo"
+import (
+	megaport "github.com/megaport/megaportgo"
+)
 
 // filterLocations filters the provided locations based on the given filters.
 func filterLocations(locations []*megaport.Location, filters map[string]string) []*megaport.Location {
@@ -20,17 +22,28 @@ func filterLocations(locations []*megaport.Location, filters map[string]string) 
 	return filtered
 }
 
-// LocationOutput represents the desired fields for JSON output.
+// LocationOutput represents the complete fields for JSON and CSV output.
 type LocationOutput struct {
-	ID        int     `json:"id"`
-	Name      string  `json:"name"`
-	Country   string  `json:"country"`
-	Metro     string  `json:"metro"`
-	SiteCode  string  `json:"site_code"`
-	Market    string  `json:"market"`
-	Latitude  float64 `json:"latitude"`
-	Longitude float64 `json:"longitude"`
-	Status    string  `json:"status"`
+	ID        int     `json:"id" header:"ID"`
+	Name      string  `json:"name" header:"Name"`
+	Country   string  `json:"country" header:"Country"`
+	Metro     string  `json:"metro" header:"Metro"`
+	SiteCode  string  `json:"site_code" header:"Site Code"`
+	Market    string  `json:"market" header:"Market"`
+	Latitude  float64 `json:"latitude" header:"-"`  // Exclude from table output
+	Longitude float64 `json:"longitude" header:"-"` // Exclude from table output
+	Status    string  `json:"status" header:"Status"`
+}
+
+// LocationTableOutput is a compact version for table display
+type LocationTableOutput struct {
+	ID       int    `header:"ID"`
+	Name     string `header:"Name"`
+	Country  string `header:"Country"`
+	Metro    string `header:"Metro"`
+	SiteCode string `header:"Site Code"`
+	Market   string `header:"Market"`
+	Status   string `header:"Status"`
 }
 
 // ToLocationOutput converts a Location to a LocationOutput.
@@ -48,8 +61,31 @@ func ToLocationOutput(l *megaport.Location) LocationOutput {
 	}
 }
 
+// ToLocationTableOutput converts a Location to a LocationTableOutput.
+func ToLocationTableOutput(l *megaport.Location) LocationTableOutput {
+	return LocationTableOutput{
+		ID:       l.ID,
+		Name:     l.Name,
+		Country:  l.Country,
+		Metro:    l.Metro,
+		SiteCode: l.SiteCode,
+		Market:   l.Market,
+		Status:   l.Status,
+	}
+}
+
 // printLocations prints the locations in the specified output format.
 func printLocations(locations []*megaport.Location, format string) error {
+	// For table format, use the compact version
+	if format == formatTable {
+		tableOutputs := make([]LocationTableOutput, 0, len(locations))
+		for _, loc := range locations {
+			tableOutputs = append(tableOutputs, ToLocationTableOutput(loc))
+		}
+		return printOutput(tableOutputs, format)
+	}
+
+	// For JSON and CSV formats, use the full output
 	outputs := make([]LocationOutput, 0, len(locations))
 	for _, loc := range locations {
 		outputs = append(outputs, ToLocationOutput(loc))
