+++ internal/cmd/output_test.go
